requirements.txt
discord.py
python-dotenv
schedule
requests